pyproject.toml
[project]
name = "metabocoach"
version = "1.0.0"
description = "Personal Metabolic AI Agent — monitors glucose, activity, and food intake to optimize health"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "google-genai>=1.0.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "twilio>=9.0.0",
    "python-telegram-bot>=21.0",
    "boto3>=1.35.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "prometheus-client>=0.21.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-secret-manager>=2.21.0",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
